+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.1.0 - 2023-06-24
+
+### Added
+- Main Features
+  1. Touch swife execution. 
+
+### Fixed
+- Corrected the demo.gif file.
+- Upgrade gradle version to 8.0.0.
+- Apply async of touch and accessibilityService permission request.
+
 ## 1.0.1 - 2023-06-19
 
 ### Fixed